+++ src/index.ts
@@ -3,6 +3,9 @@ export { SearchRoot } from "./components/SearchRoot.js";
 export { SearchInput } from "./components/SearchInput.js";
 export { SearchResults } from "./components/SearchResults.js";
 export { SearchResult } from "./components/SearchResult.js";
+export { SearchResultThumbnail } from "./components/SearchResultThumbnail.js";
+export { SearchIcon } from "./components/SearchIcon.js";
+export { SearchKbd } from "./components/SearchKbd.js";
 export { Search } from "./Search.js";
 
 // Hooks
@@ -15,5 +18,8 @@ export type { SearchRootProps } from "./components/SearchRoot.js";
 export type { SearchInputProps } from "./components/SearchInput.js";
 export type { SearchResultsProps } from "./components/SearchResults.js";
 export type { SearchResultProps } from "./components/SearchResult.js";
+export type { SearchResultThumbnailProps } from "./components/SearchResultThumbnail.js";
+export type { SearchIconProps } from "./components/SearchIcon.js";
+export type { SearchKbdProps } from "./components/SearchKbd.js";
 export type { SearchProps } from "./Search.js";
 export type { UseSearchOptions, UseSearchReturn } from "./hooks/useSearch.js";